Updated data/process_data.py to be compliant with PEP8 style. Added grid search and parameters to models/train_classifier.py.

+++ data/process_data.py
@@ -6,9 +6,9 @@ from sqlalchemy import create_engine
 def load_data(messages_filepath, categories_filepath):
     """Load in the messages and categories data sets"""
 
-    df_messages     = pandas.read_csv(messages_filepath)
-    df_categories   = pandas.read_csv(categories_filepath)
-    df              = df_messages.join(df_categories.set_index("id"), on="id")
+    df_messages = pandas.read_csv(messages_filepath)
+    df_categories = pandas.read_csv(categories_filepath)
+    df = df_messages.join(df_categories.set_index("id"), on="id")
     
     return df
 
@@ -17,15 +17,16 @@ def clean_data(df):
     """Clean the messages and categories dataframe"""
     
     # Rename the category columns and extract the values
-    categories          = df.categories.str.split(";", expand=True)
-    category_names      = categories.iloc[0].str.split("-", expand=True)[0].values
-    categories.columns  = category_names
-    categories          = categories.apply(lambda q: q.str.split("-").str[-1].astype(int), axis=0)
+    categories = df.categories.str.split(";", expand=True)
+    category_names = categories.iloc[0].str.split("-", expand=True)[0].values
+    categories.columns = category_names
+    categories = categories.apply(lambda q: q.str.split("-").str[-1], axis=0)
+    categories = categories.astype(int)
 
     # Drop the categories and join the cleaned data
     df.drop(columns=["categories"], inplace=True)
-    df  = df.join(categories)
-    df  = df.drop_duplicates()
+    df = df.join(categories)
+    df = df.drop_duplicates()
 
     return df
 
@@ -33,7 +34,7 @@ def clean_data(df):
 def save_data(df, database_filename):
     """Save the dataframe to the database"""
 
-    engine  = create_engine("sqlite:///{}".format(database_filename))
+    engine = create_engine("sqlite:///{}".format(database_filename))
     df.to_sql("Message", engine, index=False, if_exists="replace")
 
 
@@ -47,10 +48,10 @@ def main():
     parser.add_argument("database_filepath", help="File path to the database for storage")
 
     # Parse the arguments
-    args                = parser.parse_args()
-    messages_filepath   = args.messages_filepath
+    args = parser.parse_args()
+    messages_filepath = args.messages_filepath
     categories_filepath = args.categories_filepath
-    database_filepath   = args.database_filepath
+    database_filepath = args.database_filepath
 
     # Process the ETL pipeline
     print('Loading data...\n    MESSAGES: {}\n    CATEGORIES: {}'
